More lazily evaluated methods: `map`, `filter`, `repeat`, `shuffle`
