+++ lhotse/dataset/sampling/__init__.py
@@ -1,6 +1,5 @@
 from .bucketing import BucketingSampler
 from .cut_pairs import CutPairsSampler
-from .data_source import streaming_shuffle
 from .dynamic import DynamicCutSampler
 from .dynamic_bucketing import DynamicBucketingSampler
 from .simple import SimpleCutSampler, SingleCutSampler
