+++ lhotse/utils.py
@@ -14,6 +14,7 @@ from typing import (
     Any,
     Callable,
     Dict,
+    Generator,
     Iterable,
     List,
     Optional,
@@ -706,3 +707,49 @@ class suppress_and_warn:
                 f"[Suppressed {exctype.__qualname__}] Error message: {excinst}"
             )
         return should_suppress
+
+
+def streaming_shuffle(
+    data: Iterable[T],
+    bufsize: int = 10000,
+    rng: Optional[random.Random] = None,
+) -> Generator[T, None, None]:
+    """
+    Shuffle the data in the stream.
+
+    This uses a buffer of size ``bufsize``. Shuffling at
+    startup is less random; this is traded off against
+    yielding samples quickly.
+
+    This code is mostly borrowed from WebDataset; note that we use much larger default
+    buffer size because Cuts are very lightweight and fast to read.
+    https://github.com/webdataset/webdataset/blob/master/webdataset/iterators.py#L145
+
+    .. warning: The order of the elements is expected to be much less random than
+        if the whole sequence was shuffled before-hand with standard methods like
+        ``random.shuffle``.
+
+    :param data: iterator
+    :param bufsize: buffer size for shuffling
+    :param rng: either random module or random.Random instance
+    :return: a generator of cuts, shuffled on-the-fly.
+    """
+    if rng is None:
+        rng = random
+    buf = []
+    startup = True
+    for sample in data:
+        if len(buf) < bufsize:
+            try:
+                buf.append(next(data))
+            except StopIteration:
+                pass
+        k = rng.randint(0, len(buf) - 1)
+        sample, buf[k] = buf[k], sample
+        if startup and len(buf) < bufsize:
+            buf.append(sample)
+            continue
+        startup = False
+        yield sample
+    for sample in buf:
+        yield sample
