+++ test/test_manipulation.py
@@ -133,19 +133,6 @@ def test_combine(manifest_type):
 
 
 @mark.parametrize("manifest_type", [RecordingSet, SupervisionSet, FeatureSet, CutSet])
-def test_combine_lazy(manifest_type):
-    expected = DummyManifest(manifest_type, begin_id=0, end_id=200)
-    with as_lazy(DummyManifest(manifest_type, begin_id=0, end_id=68)) as part1, as_lazy(
-        DummyManifest(manifest_type, begin_id=68, end_id=136)
-    ) as part2, as_lazy(
-        DummyManifest(manifest_type, begin_id=136, end_id=200)
-    ) as part3:
-        combined = combine(part1, part2, part3)
-        # Equivalent under iteration
-        assert list(combined) == list(expected)
-
-
-@mark.parametrize("manifest_type", [RecordingSet, SupervisionSet, FeatureSet, CutSet])
 def test_subset_first(manifest_type):
     any_set = DummyManifest(manifest_type, begin_id=0, end_id=200)
     expected = DummyManifest(manifest_type, begin_id=0, end_id=10)
